moved parse_params from util to settings
